refactor gestureOptions to one source of truth
