+++ src/app/(signed-in)/_trace/trace/[traceId]/edit/components/repair-screen/repair-screen.tsx
@@ -3,20 +3,7 @@
 import React, { useCallback, useEffect, useState } from "react";
 import Image from "next/image";
 import {
-  ArrowDownFromLine,
-  ArrowLeftFromLine,
-  ArrowRightFromLine,
-  ArrowUpFromLine,
-  Circle,
   CircleAlert,
-  CircleDot,
-  CircleHelp,
-  CircleStop,
-  Expand,
-  Grab,
-  IterationCcw,
-  IterationCw,
-  Shrink,
 } from "lucide-react";
 
 import {
@@ -31,104 +18,7 @@ import { cn } from "@/lib/utils";
 import { useFormContext, useWatch } from "react-hook-form";
 import { TraceFormData } from "../types";
 import { useHotkeys } from "react-hotkeys-hook";
-
-export const gestureOptions = [
-  {
-    value: "tap",
-    label: "Tap",
-    icon: <Circle className="size-4 text-yellow-800 hover:text-black" />,
-  },
-  {
-    value: "double tap",
-    label: "Double tap",
-    icon: <CircleDot className="size-4 text-yellow-800 hover:text-black" />,
-  },
-  {
-    value: "touch and hold",
-    label: "Touch and hold",
-    icon: <CircleStop className="size-4 text-yellow-800 hover:text-black" />,
-  },
-  {
-    value: "swipe",
-    label: "Swipe",
-    subGestures: [
-      {
-        value: "swipe up",
-        label: "Swipe up",
-        icon: (
-          <ArrowUpFromLine className="size-4 text-yellow-800 hover:text-black" />
-        ),
-      },
-      {
-        value: "swipe down",
-        label: "Swipe down",
-        icon: (
-          <ArrowDownFromLine className="size-4 text-yellow-800 hover:text-black" />
-        ),
-      },
-      {
-        value: "swipe left",
-        label: "Swipe left",
-        icon: (
-          <ArrowLeftFromLine className="size-4 text-yellow-800 hover:text-black" />
-        ),
-      },
-      {
-        value: "swipe right",
-        label: "Swipe right",
-        icon: (
-          <ArrowRightFromLine className="size-4 text-yellow-800 hover:text-black" />
-        ),
-      },
-    ],
-  },
-  {
-    value: "drag",
-    label: "Drag",
-    icon: <Grab className="size-4 text-yellow-800 hover:text-black" />,
-  },
-  {
-    value: "zoom",
-    label: "Zoom",
-    subGestures: [
-      {
-        value: "zoom in",
-        label: "Zoom in",
-        icon: <Shrink className="size-4 text-yellow-800 hover:text-black" />,
-      },
-      {
-        value: "zoom out",
-        label: "Zoom out",
-        icon: <Expand className="size-4 text-yellow-800 hover:text-black" />,
-      },
-    ],
-  },
-  {
-    value: "rotate",
-    label: "Rotate",
-    subGestures: [
-      {
-        value: "rotate cw",
-        label: "Rotate clockwise",
-        icon: (
-          <IterationCw className="size-4 text-yellow-800 hover:text-black" />
-        ),
-      },
-      {
-        value: "rotate ccw",
-        label: "Rotate counter-clockwise",
-        icon: (
-          <IterationCcw className="size-4 text-yellow-800 hover:text-black" />
-        ),
-      },
-    ],
-  },
-  {
-    value: "other",
-    label: "Other",
-    icon: <CircleHelp className="size-4 text-yellow-800 hover:text-black" />,
-  },
-];
+import { gestureOptions } from "@/lib/utils/gesture-options";
 
 export default function RepairScreen({ trace }: { trace: any }) {
   const [watchScreens, watchVHs, watchGestures] = useWatch({
