+++ src/app/(signed-in)/_trace/trace/[traceId]/edit/components/review/review.tsx
@@ -8,7 +8,7 @@ import { useFormContext } from "react-hook-form";
 import { TraceFormData } from "../types";
 import { Screen, ScreenGesture } from "@prisma/client";
 import { Tooltip, TooltipTrigger, TooltipContent, TooltipProvider } from "@/components/ui/tooltip";
-import { gestureOptions } from "../repair-screen";
+import { gestureOptions } from "@/lib/utils/gesture-options";
 
 export default function Review() {
   const { register } = useFormContext<TraceFormData>();
