+++ src/app/(signed-in)/_trace/trace/[traceId]/edit/components/types.ts
@@ -67,24 +67,6 @@ export const GestureSchema = z.record(
   })
 );
 
-export const GestureOptionSchema: z.ZodType<{
-  value: string;
-  label: string;
-  icon?: React.JSX.Element;
-  subGestures?: any;
-}> = z.lazy(
-  (): z.ZodType<any> => // 👈 annotate the return type here
-    z.object({
-      value: z.string(),
-      label: z.string(),
-      icon: z.custom<React.JSX.Element>().optional(),
-      subGestures: z.array(GestureOptionSchema).optional(),
-    })
-);
-
-// Then define the type from schema (for safety & completion support)
-export type GestureOption = z.infer<typeof GestureOptionSchema>;
-
 export const ScreenGestureSchema = z
   .object({
     screens: ScreenSchema,
